Highlight excessively long lists of content
In order to help curate the length of lists down to a user-friendly
size, we want a view to present to publishers to help illustrate
parts of the taxonomy with a lot of content.

This is achieved by letting users visit the prototype with a special
query string in their URL: `?longLists`. This “mode” stays enabled
with a cookie, until the user disables it with `?longLists=false`.

When this mode is enabled, they will be able to see accordion sections
which are too long. Expanding the section will have a line ruled across
the content limit, telling the user that content beyond this line is
rarely clicked.